+++ shared/NaviMeshHandler.cpp
@@ -35,7 +35,7 @@
 static const int NAVMESHSET_MAGIC = 'M'<<24 | 'S'<<16 | 'E'<<8 | 'T'; //'MSET';
 static const int NAVMESHSET_VERSION = 1;
 static const int MAX_AGENTS = 128;
-static const int _MAX_AGENT_RADIUS_ = 3;
+static const float _MAX_AGENT_RADIUS_ = 3;
 
 
 struct NavMeshSetHeader
